# #27 Handle copying referenced objects