--- a/package.json
+++ b/package.json
@@ -1,5 +1,6 @@
 {
   "author": "tony_quetano@planttheidea.com",
+  "contributors": ["Dariusz Rzepka <rzepkadarek@gmail.com>"],
   "browser": "dist/fast-copy.js",
   "bugs": {
     "url": "https://github.com/planttheidea/fast-copy/issues"