Solves #27 

Regarding using WeakMap - can we require to use a polyfill for that or we want to implement some simple implementation of it? It's widely supported across different browsers - https://caniuse.com/#search=weakmap


Benchmark - it's slightly slower with this fix. Maybe we can improve something? Or is simply WeakMap that much slower than the WeakSet?
```
Starting benchmarks for group simple object...
Master:
Benchmark completed for fast-copy: 2,258,798 ops/sec
Benchmark completed for fast-copy (strict): 514,319 ops/sec
After:
Benchmark completed for fast-copy: 2,235,410 ops/sec
Benchmark completed for fast-copy (strict): 479,978 ops/sec
```
```
Starting benchmarks for group complex object...
Master:
Benchmark completed for fast-copy: 100,285 ops/sec
Benchmark completed for fast-copy (strict): 34,349 ops/sec
After:
Benchmark completed for fast-copy: 82,292 ops/sec
Benchmark completed for fast-copy (strict): 30,649 ops/sec
```
```
Starting benchmarks for group circular object...
Master:
Benchmark completed for fast-copy: 1,200,336 ops/sec
Benchmark completed for fast-copy (strict): 372,298 ops/sec
After:
Benchmark completed for fast-copy: 1,193,205 ops/sec
Benchmark completed for fast-copy (strict): 354,410 ops/sec
```

```
Benchmark results complete, overall averages:
Master:
┌────────────────────┬───────────┐
│ Name               │ Ops / sec │
├────────────────────┼───────────┤
│ fast-copy          │ 667.553   │
├────────────────────┼───────────┤
│ lodash.cloneDeep   │ 506.869   │
├────────────────────┼───────────┤
│ clone              │ 445.201   │
├────────────────────┼───────────┤
│ ramda              │ 344.407   │
├────────────────────┼───────────┤
│ fast-deepclone     │ 278.045   │
├────────────────────┼───────────┤
│ deepclone          │ 250.351   │
├────────────────────┼───────────┤
│ fast-copy (strict) │ 204.973   │
├────────────────────┼───────────┤
│ fast-clone         │ 162.576   │
└────────────────────┴───────────┘
After:
┌────────────────────┬───────────┐
│ Name               │ Ops / sec │
├────────────────────┼───────────┤
│ fast-copy          │ 718.736   │
├────────────────────┼───────────┤
│ lodash.cloneDeep   │ 520.245   │
├────────────────────┼───────────┤
│ clone              │ 435.118   │
├────────────────────┼───────────┤
│ ramda              │ 390.288   │
├────────────────────┼───────────┤
│ fast-deepclone     │ 325.057   │
├────────────────────┼───────────┤
│ deepclone          │ 262.127   │
├────────────────────┼───────────┤
│ fast-clone         │ 172.48    │
├────────────────────┼───────────┤
│ fast-copy (strict) │ 155.235   │
└────────────────────┴───────────┘
```