@darkowic 2020-02-25
This benchmark is not reliable at all... I get different results all the time - but constantly the old version is slightly faster