@planttheidea 2020-02-29
> can we require to use a polyfill for that or we want to implement some simple implementation of it? It's widely supported across different browsers - https://caniuse.com/#search=weakmap

Suddenly requiring this would be a breaking change, and I would like to avoid this as much as possible for what really is a bug fix. There is already an existing fallback implementation, and tweaking it for the same results would be trivial.

> This benchmark is not reliable at all

Welcome to JavaScript benchmarks. 😋 The goal is to see themes, not treat the results as biblical (it varies from machine to machine, and even on the same machine it varies from run to run due to system load and power management). That said, slower here does not mean bad, because better accuracy is more important, especially since the reduction is small.

I'm on vacation right now, but once I'm back I can investigate more thoroughly. Looks solid at first glance, though, once the fallback is in place.