@darkowic 2020-03-01
Thank you for the answer @planttheidea. I have added a simple WeakMap polyfill implementation. Should be enough as a simple fallback.

About the benchmark: I totally agree - it is slightly slower but still keeps the first place and solves the issue correctly.